requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Security
